refactor: create apps dir and move here some stuff
