+++ README.md
@@ -74,7 +74,7 @@ This project is based on Django, so all usual Django rules and commands are vali
     For example, this sequence of commands will make you create a dummy survey:
     -   import service:
         ```
-        from polls.services.poll_service import PollService
+        from apps.polls_management.services.poll_service import PollService
         ```
     -   use service to create dummy survey
         ```
