+++ apps/polls_management/classes/majority_poll_result_data.py
@@ -1,6 +1,6 @@
 from polls.models.majority_judgment_model import MajorityJudgmentModel
 from polls.models.poll_option_model import PollOptionModel
-from polls.exceptions.poll_not_yet_voted_exception import PollNotYetVodedException
+from apps.polls_management.exceptions.poll_not_yet_voted_exception import PollNotYetVodedException
 
 from dataclasses import dataclass
 import math
