+++ apps/polls_management/classes/poll_result.py
