+++ apps/polls_management/constants/models_constants.py
