+++ apps/polls_management/exceptions/majority_number_of_ratings_not_valid.py
