+++ apps/polls_management/exceptions/poll_not_yet_voted_exception.py
