+++ apps/polls_management/services/majority_vote_service.py
@@ -1,10 +1,10 @@
 import math
 from typing import List
 from django.core.exceptions import ObjectDoesNotExist
-from polls.classes.majority_poll_result_data import MajorityPollResultData
-from polls.exceptions.majority_number_of_ratings_not_valid import MajorityNumberOfRatingsNotValid
-from polls.exceptions.poll_does_not_exist_exception import PollDoesNotExistException
-from polls.exceptions.poll_option_rating_unvalid_exception import PollOptionRatingUnvalidException
+from apps.polls_management.classes.majority_poll_result_data import MajorityPollResultData
+from apps.polls_management.exceptions.majority_number_of_ratings_not_valid import MajorityNumberOfRatingsNotValid
+from apps.polls_management.exceptions.poll_does_not_exist_exception import PollDoesNotExistException
+from apps.polls_management.exceptions.poll_option_rating_unvalid_exception import PollOptionRatingUnvalidException
 from polls.models.majority_judgment_model import MajorityJudgmentModel
 from polls.models.majority_vote_model import MajorityVoteModel
 from polls.models.poll_model import PollModel
