+++ apps/polls_management/services/poll_create_service.py
@@ -1,8 +1,8 @@
 from typing import List
-from polls.exceptions.poll_not_valid_creation_exception import *
-from polls.models import PollModel
+from apps.polls_management.exceptions.poll_not_valid_creation_exception import *
+from polls.models.poll_model import PollModel
 from polls.models.poll_option_model import PollOptionModel
-from polls.classes.poll_form import PollForm
+from apps.polls_management.classes.poll_form import PollForm
 
 
 class PollCreateService:
