+++ apps/polls_management/services/vote_service.py
@@ -1,11 +1,11 @@
-from polls.classes.vote_builder import VoteBuilder
+from apps.polls_management.classes.vote_builder import VoteBuilder
 from polls.models.poll_model import PollModel
 from polls.models.vote_model import VoteModel
-from polls.exceptions.poll_does_not_exist_exception import PollDoesNotExistException
-# from polls.exceptions.poll_option_unvalid_exception import PollOptionUnvalidException
-from polls.exceptions.vote_does_not_exixt_exception import VoteDoesNotExistException
+from apps.polls_management.exceptions.poll_does_not_exist_exception import PollDoesNotExistException
+# from apps.polls_management.exceptions.poll_option_unvalid_exception import PollOptionUnvalidException
+from apps.polls_management.exceptions.vote_does_not_exixt_exception import VoteDoesNotExistException
 from django.core.exceptions import ObjectDoesNotExist
-from polls.classes.poll_result import PollResult
+from apps.polls_management.classes.poll_result import PollResult
 
 class VoteService: 
     """
