+++ apps/polls_management/views/majority_view.py
@@ -3,14 +3,14 @@ from django.http import Http404
 from django.http import HttpRequest, HttpResponseServerError, HttpResponseRedirect
 from django.shortcuts import render
 from django.urls import reverse
-from polls.classes.majority_poll_result_data import MajorityPollResultData
-from polls.exceptions.poll_does_not_exist_exception import PollDoesNotExistException
-from polls.exceptions.poll_not_yet_voted_exception import PollNotYetVodedException
-from polls.exceptions.poll_option_rating_unvalid_exception import PollOptionRatingUnvalidException
+from apps.polls_management.classes.majority_poll_result_data import MajorityPollResultData
+from apps.polls_management.exceptions.poll_does_not_exist_exception import PollDoesNotExistException
+from apps.polls_management.exceptions.poll_not_yet_voted_exception import PollNotYetVodedException
+from apps.polls_management.exceptions.poll_option_rating_unvalid_exception import PollOptionRatingUnvalidException
 from polls.models.majority_vote_model import MajorityVoteModel
 from polls.models.poll_model import PollModel
-from polls.services.majority_vote_service import MajorityVoteService
-from polls.services.poll_service import PollService
+from apps.polls_management.services.majority_vote_service import MajorityVoteService
+from apps.polls_management.services.poll_service import PollService
 
 
 def dummy_majority(request: HttpRequest, poll_id=None):
