+++ documentation/code_documentation/polls/services/poll_create_service.md
@@ -1,2 +1,2 @@
 # Poll create service
-::: polls.services.poll_create_service
+::: apps.polls_management.services.poll_create_service
