+++ polls/models/majority_vote_model.py
@@ -1,7 +1,7 @@
 from typing import List
 
 from polls.models.majority_judgment_model import MajorityJudgmentModel
-from ..constants.models_constants import POLL_MODEL_NAME
+from apps.polls_management.constants.models_constants import POLL_MODEL_NAME
 from django.db import models
 
 class MajorityVoteModel(models.Model):
