+++ test/polls/test_vote_service.py
@@ -1,13 +1,13 @@
-from polls.classes.poll_result import PollResult, PollResultVoice
+from apps.polls_management.classes.poll_result import PollResult, PollResultVoice
 from polls.models.vote_model import VoteModel
 import pytest
 from assertpy import assert_that
 from polls.models.poll_model import PollModel
 from polls.models.poll_option_model import PollOptionModel
-from polls.services.vote_service import VoteService
-from polls.exceptions.poll_option_unvalid_exception import PollOptionUnvalidException
-from polls.exceptions.poll_does_not_exist_exception import PollDoesNotExistException
-from polls.exceptions.vote_does_not_exixt_exception import VoteDoesNotExistException
+from apps.polls_management.services.vote_service import VoteService
+from apps.polls_management.exceptions.poll_option_unvalid_exception import PollOptionUnvalidException
+from apps.polls_management.exceptions.poll_does_not_exist_exception import PollDoesNotExistException
+from apps.polls_management.exceptions.vote_does_not_exixt_exception import VoteDoesNotExistException
 
 @pytest.fixture()
 def test_polls(request):
